+++ netqasm/lang/instr/flavour.py
@@ -117,6 +117,15 @@ class NVFlavour(Flavour):
         super().__init__(self.instrs)
 
 
+class REIDSFlavour(Flavour):
+    @property
+    def instrs(self):
+        return []
+
+    def __init__(self):
+        super().__init__(self.instrs)
+
+
 class TrappedIonFlavour(Flavour):
     @property
     def instrs(self):