+++ netqasm/runtime/interface/config.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 from enum import Enum
-from typing import Any, Dict, List
+from typing import Any, Dict, List, Optional
 
 from netqasm.util.yaml import load_yaml
 
@@ -125,13 +125,12 @@ def parse_network_config(cfg) -> NetworkConfig:
 # Keys are role names, values are node names.
 RolesConfig = Dict[str, str]
 
-
 # App input per role.
 # Keys are names of variables that are passed to the role's `main` function.
 AppInput = Dict[str, Any]
 
 
-def network_cfg_from_file(network_config_file: str = None) -> NetworkConfig:
+def network_cfg_from_file(network_config_file: Optional[str] = None) -> NetworkConfig:
     yaml_dict = load_yaml(network_config_file)
     network_cfg = parse_network_config(yaml_dict)
     return network_cfg