+++ netqasm/sdk/classical_communication/thread_socket/socket.py
@@ -401,7 +401,10 @@ class ThreadSocket(Socket):
         self._SOCKET_HUB.send(self, msg)
 
     def recv_silent(
-        self, block: bool = True, timeout: Optional[float] = None, maxsize: int = None
+        self,
+        block: bool = True,
+        timeout: Optional[float] = None,
+        maxsize: Optional[int] = None,
     ) -> str:
         """Receive a message without logging"""
         msg = self._SOCKET_HUB.recv(self, block=block, timeout=timeout)